pyproject.toml
[project]
name = "reel-downloader"
version = "0.1.0"
description = "Telegram bot to download instagram reels"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.18.0",
    "pytest-playwright>=0.7.0",
]
